requirements.txt
# 知识库系统依赖包

# 核心依赖
redis>=4.5.0
numpy>=1.24.0
sentence-transformers>=2.2.0

# 文件解析
python-docx>=0.8.11
pywin32>=306
olefile>=0.47
openpyxl>=3.1.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
Pillow>=9.5.0
pytesseract>=0.3.10
rapidocr_onnxruntime>=1.4.0
easyocr>=1.7.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web服务
flask>=2.3.0
flask-cors>=4.0.0

# 剪贴板
pyperclip>=1.8.2

# 智谱AI
zhipuai>=2.0.0

# 可选依赖
watchdog>=3.0.0
